Revert "CRSF with Angular - METHOD 3 (angular httpProvider configuration)"
This reverts commit f56358cb4d369cdafb3d384e3efa3107b9a11c58.
Revert to using server side crsf fix. Adjusting the headers client side (with angular) caused problems with cross origin policy and CORS when accessing 3rd party apis with javascript. The presence of additional header with a crsf token was not expected by the 3rd party api (yahoo in this case) and so the request failed.

Conflicts:
	app/assets/javascripts/main.js